Changed NLPF and main
Separate three inputs:
1) starting point
2) guess for the solution
3) fundamental changes.

In terms of I-O, I change the saving of the output to outside the NLPF function. This avoids having to have an indicator for SS in the function (which becomes troublesome when SS can take more than 0 or 1).

In the new function, the only three things that matter for this function is: starting point, shock (in time difference), and guesses.

SS versus non SS are but two special cases with different time horizons (controlled via hat_fundamentals).

The main is updated accordingly.  In particular, I tried not to have too many "loose" variables around that could be used in calling NLPF_HAT, in that most variables are wrapped in a struture and not unpacked unless the variable is a global constant and is always the same across the entire NLPF.

In the current code, the nonlinear NLPF is called on the base of the SS.